Update profile design
