--- packages/ui/src/components/HorizontalTabs/index.ts
@@ -1,1 +1,0 @@
-export * from './HorizontalTabs'
