+++ packages/ui/src/components/ProfileSettings/ProfilePanel.tsx
@@ -1,7 +1,16 @@
 'use client'
 import { FC, PropsWithChildren } from 'react'
 
-import { Center, HStack, Stack, Text, VStack } from '@chakra-ui/react'
+import {
+  Box,
+  Tab,
+  TabList,
+  TabPanel,
+  TabPanels,
+  Tabs,
+  Text,
+  VStack,
+} from '@chakra-ui/react'
 import { useTranslation } from 'next-i18next'
 import { FaPaintBrush, FaUserCircle } from 'react-icons/fa'
 import { FaKey } from 'react-icons/fa6'
@@ -9,13 +18,12 @@ import { TbSocial } from 'react-icons/tb'
 
 import { useAuthContext } from '@fc/context'
 
-import { Arts } from './AppSpecificTabs/Art'
-import { Details } from './CommonTabs/Details'
-import { Security } from './CommonTabs/Security'
-import { Socials } from './CommonTabs/Socials'
+import { ArtsTab } from './ArtsTab'
+import { DetailsTab } from './DetailsTab'
+import { SecurityTab } from './SecurityTab'
+import { Socials } from './SocialsTab'
 import { Container } from '../Container'
 import { Hero } from '../Hero'
-import { HorizontalTabs, HorizontalTab } from '../HorizontalTabs'
 import { WAvatar } from '../WAvatar'
 
 type ProfilePanelProps = PropsWithChildren<{
@@ -42,56 +50,63 @@ export const ProfilePanel: FC<ProfilePanelProps> = ({
             name={profile?.name || user?.username}
             boxShadow={'2xl'}
           />
-          <HStack
-            justifyContent="center"
-            alignItems={'center'}
-            alignContent={'flex-end'}
-            bg="transparent"
+          <Box
+            color={'white'}
+            fontSize={'lg'}
+            textTransform={'capitalize'}
+            textAlign={'center'}
+            fontWeight={600}
           >
-            <Text color={'white'}>{profile?.name || user?.username}</Text>
-          </HStack>
+            <Text>{profile?.name || user?.username}</Text>
+            <Text>{user.roles.join(', ')}</Text>
+          </Box>
         </VStack>
-        <Stack
-          position={'absolute'}
-          right={2}
-          bottom={2}
-          spacing={4}
-          align={'flex-end'}
-        >
-          <Text color={'white'}>Role: {user.roles.join(', ')}</Text>
-        </Stack>
       </Hero>
 
-      <Container>
-        <Center>
-          <HorizontalTabs>
-            <HorizontalTab
-              title={t('profile.tabs.profile')}
-              icon={<FaUserCircle />}
-            >
-              <Details />
-            </HorizontalTab>
-            <HorizontalTab title={t('profile.tabs.security')} icon={<FaKey />}>
-              <Security />
-            </HorizontalTab>
-            <HorizontalTab
-              title={t('profile.tabs.socials')}
-              icon={<TbSocial />}
-            >
-              <Socials />
-            </HorizontalTab>
-            {showArts && (
-              <HorizontalTab
-                title={t('profile.tabs.arts')}
-                icon={<FaPaintBrush />}
-              >
-                <Arts />
-              </HorizontalTab>
-            )}
+      <Box bg="white" p={8}>
+        <Container>
+          <Tabs
+            orientation="vertical"
+            border={0}
+            colorScheme="primary"
+            size="lg"
+          >
+            <TabList border={0} minW={300} h="max-content">
+              <Tab fontWeight={600} justifyContent={'start'}>
+                <Box as={FaUserCircle} mr={2} /> {t('profile.tabs.profile')}
+              </Tab>
+              <Tab fontWeight={600} justifyContent={'start'}>
+                <Box as={FaKey} mr={2} /> {t('profile.tabs.security')}
+              </Tab>
+              <Tab fontWeight={600} justifyContent={'start'}>
+                <Box as={TbSocial} mr={2} /> {t('profile.tabs.socials')}
+              </Tab>
+              {showArts && (
+                <Tab fontWeight={600} justifyContent={'start'}>
+                  <Box as={FaPaintBrush} mr={2} /> {t('profile.tabs.arts')}
+                </Tab>
+              )}
+            </TabList>
+            <TabPanels>
+              <TabPanel>
+                <DetailsTab />
+              </TabPanel>
+              <TabPanel>
+                <SecurityTab />
+              </TabPanel>
+              <TabPanel>
+                <Socials />
+              </TabPanel>
+              {showArts && (
+                <TabPanel>
+                  <ArtsTab />
+                </TabPanel>
+              )}
+            </TabPanels>
             {children}
-          </HorizontalTabs>
-        </Center>
-      </Container>
+          </Tabs>
+        </Container>
+      </Box>
     </>
   )
 }
