+++ packages/ui/src/components/index.ts
@@ -33,7 +33,6 @@ export * from './FormItem'
 export * from './FormUploader'
 export * from './FormattedDate'
 export * from './Header'
-export * from './HorizontalTabs'
 export * from './Hero'
 export * from './Icons'
 export * from './JoinForm'
@@ -55,13 +54,13 @@ export * from './PostSentenceForm'
 export * from './Profile'
 export * from './ProfileSettings'
 export * from './ResetPasswordForm'
-export * from './TranslateLocales'
 export * from './SearchForm'
 export * from './ShareButtons'
 export * from './SignupForm'
 export * from './SocialButtons'
 export * from './SocialLoginButtons'
 export * from './SubToNotification'
+export * from './TranslateLocales'
 export * from './UserFeedback'
 export * from './UserFeedbackForm'
 export * from './VideoPlayer'
